fix: change scrollElement when navigate
